+++ CHANGELOG.md
@@ -4,7 +4,7 @@ CHANGELOG
 [Next release](https://github.com/rebing/graphql-laravel/compare/2.1.0...master)
 --------------
 ### Changed
-- Forward PHP engine errors to the application error handler
+- Forward PHP engine errors to the application error handler [\487 / mfn](https://github.com/rebing/graphql-laravel/pull/487)
 
 2019-08-27, 2.1.0
 -----------------